+++ android/src/main/java/com/bleplx/converter/DescriptorToJsObjectConverter.java
@@ -29,10 +29,6 @@ public class DescriptorToJsObjectConverter extends JSObjectConverter<Descriptor>
     js.putInt(Metadata.SERVICE_ID, descriptor.getServiceId());
     js.putString(Metadata.SERVICE_UUID, UUIDConverter.fromUUID(descriptor.getServiceUuid()));
     js.putString(Metadata.DEVICE_ID, descriptor.getDeviceId());
-
-    if (descriptor.getValue() == null) {
-      descriptor.setValueFromCache();
-    }
     js.putString(Metadata.VALUE, descriptor.getValue() != null ? Base64Converter.encode(descriptor.getValue()) : null);
     return js;
   }