+++ example-expo/src/utils/cloneDeep.ts
@@ -1,1 +1,0 @@
-export const cloneDeep: <T>(objectToClone: T) => T = objectToClone => JSON.parse(JSON.stringify(objectToClone))