+++ example/src/theme/colors.ts
@@ -1,3 +1,0 @@
-export const colors = {
-  mainRed: '#ff304d'
-} as const