+++ src/BleManager.js
@@ -391,6 +391,38 @@ export class BleManager {
     return this._callPromise(BleModule.stopDeviceScan())
   }
 
+  async startBackgroundDeviceScan(UUIDs: ?Array<UUID>, options: ?Object): Promise<void> {
+    return this._callPromise(BleModule.startBackgroundDeviceScan(UUIDs, options))
+  }
+
+  stopBackgroundDeviceScan(): Promise<void> {
+    return this._callPromise(BleModule.stopBackgroundDeviceScan())
+  }
+
+  isBackgroundScanRunning(): Promise<boolean> {
+    return this._callPromise(BleModule.isBackgroundScanRunning())
+  }
+
+  async startBackgroundDataCollection(options: ?Object): Promise<void> {
+    return this._callPromise(BleModule.startBackgroundDataCollection(options))
+  }
+
+  async connectBackgroundDevice(deviceId: DeviceId, serviceUUID: UUID, characteristicUUID: UUID): Promise<void> {
+    return this._callPromise(BleModule.connectBackgroundDevice(deviceId, serviceUUID, characteristicUUID))
+  }
+
+  async stopBackgroundDataCollection(): Promise<void> {
+    return this._callPromise(BleModule.stopBackgroundDataCollection())
+  }
+
+  async getPendingBackgroundData(): Promise<Array<Object>> {
+    return this._callPromise(BleModule.getPendingBackgroundData())
+  }
+
+  async clearPendingBackgroundData(): Promise<void> {
+    return this._callPromise(BleModule.clearPendingBackgroundData())
+  }
+
   /**
    * Request a connection parameter update. This functions may update connection parameters on Android API level 21 or
    * above.